refactor into use goal page and use goal progress hooks and update components to use
