+++ src/components/goals/goal-view-page.tsx
@@ -1,10 +1,10 @@
 import GoalProgress from "./goal-progress";
 import UpdateGoal from "./update-goal";
 
-export default function GoalViewPage({ userId }: { userId: number }) {
+export default function GoalViewPage() {
   return (
     <div className="flex flex-row gap-6 m-6">
-      <GoalProgress userId={userId} />
+      <GoalProgress />
       <UpdateGoal />
     </div>
   );
